bug fixes: temps now show accuarte values

+++ Mathster_BMP280.h
@@ -31,7 +31,7 @@ private:
   uint8_t press_reg_start       = 0xF7;                                                    //till and including 0xF9;
   uint16_t dig_T1, dig_P1;                                                                 //calibration
   uint8_t calibration_reg_start = 0x88;                                                    //till and including 0x9F ,i.e. 24 bytes
-  int32_t dig_T2, dig_T3, dig_P2, dig_P3, dig_P4, dig_P5, dig_P6, dig_P7, dig_P8, dig_P9;  //calibration
+  int16_t dig_T2, dig_T3, dig_P2, dig_P3, dig_P4, dig_P5, dig_P6, dig_P7, dig_P8, dig_P9;  //calibration
   int32_t t_fine;                                                                          //internal use only
   int temp_internal;                                                                       //internal use only
   uint8_t i2c_read_byte(const uint8_t addr);
